pyproject.toml
[tool.poetry]
name = "ai-workflow-core"
version = "0.1.0"
description = ""
authors = ["evachenirev <evachenirev03@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.45.1"
openai = "^1.83.0"
python-dotenv = "^1.1.0"
pytest = "^8.4.0"
google-generativeai = "^0.8.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
